1.HIK增加USB类型相机 2.解决HIK内存占用过高问题 3.其他细节优化

+++ src/Common/Model/CamConfig.cs
@@ -4,17 +4,17 @@
     {
         public TriggerMode triggerMode { get; set; }
 
-        public TriggerSource triggeSource { get; set; } 
+        public TriggerSource triggeSource { get; set; }
 
         public TriggerPolarity triggerPolarity { get; set; }
 
-        public ushort ExpouseTime { get; set; }
+        public ulong ExpouseTime { get; set; }
 
         public ushort TriggerFilter { get; set; }
 
         public ushort TriggerDelay { get; set; }
 
-        public short Gain { get; set; }
+        public float Gain { get; set; }
     }
 }
